+++ packages/fr_storage/lib/fr_storage.dart
@@ -1,4 +1,5 @@
 library;
 
+export 'src/fr_box.dart' show FrBox;
 export 'src/fr_storage.dart' show FrStorage;
-export 'src/key_value_storage.dart';
+export 'src/fr_storage_instance.dart' show FrStorageInstance;